feat: add extra 'Appearance' settings page
